refactor(compiler-plugin/ir): split collectPreviewsReplacement logic + buildPreviewSequence sub-logic
Split the IR side of the `previewCollection` feature into its dedicated
`collectPreviewsReplacement/` logic and `buildPreviewSequence/` sub-logic.

Renames + moves:
- `ir/PreviewLabIrBodyFiller.kt`
  → `feature/previewCollection/ir/collectPreviewsReplacement/ReplaceCollectPreviewsFunBody.kt`
  (class `PreviewLabIrBodyFiller` → `ReplaceCollectPreviewsFunBody`; verb-prefix per project convention)
- `ir/HintDiscovery.kt`
  → `feature/previewCollection/ir/collectPreviewsReplacement/DiscoverHints.kt`
  (function name kept as `discoverHints` — top-level function, no class wrap)
- `ir/PreviewListIrBuilder.kt`
  → 4 files under `.../buildPreviewSequence/`:
    - `BuildPreviewSequenceIr.kt` — orchestrating builder + shared
      `PreviewSequenceBuildContext` for sub-builders
    - `BuildLazyWrapperIr.kt` — `lazy { sequenceExpr }` wrapper
    - `BuildPreviewExportIr.kt` — `PreviewExport(lazyExpr)` ctor call
    - `BuildConcatenatedPreviewSequencesIr.kt` — cross-module concat +
      `distinctPreviewsByIdSequence` for `collectAllModulePreviews()`
- `ir/CollectedPreviewIrBuilder.kt`
  → `feature/previewCollection/ir/collectPreviewsReplacement/buildPreviewSequence/BuildCollectedPreviewIr.kt`
  (class renamed; `buildCollectedPreviewCall` → `operator fun invoke`)
- `ir/ExtractedSourceText.kt`
  → `feature/previewCollection/ir/collectPreviewsReplacement/buildPreviewSequence/`
  (data class kept; only the package moved)

Splits (1 → 3 files):
- `ir/PreviewHintIrBodyFiller.kt` →
    - `FillPreviewHintIrBody.kt` — visitor + body filling logic
    - `BuildPreviewByHashMap.kt` — `class BuildPreviewByHashMap { operator fun invoke(...) }`
      that builds the canonical-key-hashed map for the body filler
    - `HashMapWithCollisionDetection.kt` — collision-detecting map builder
      with explicit `intentionally specialized to preview hint hashing — do
      not generalize without moving to utils/` notice in its KDoc

The composition shape is the same — `PreviewLabIrGenerationExtension`
constructs the renamed classes in the same order — but each class now lives
in its own file with a `Build<...>Ir { operator fun invoke(...): IrExpression }`
shape. The shared lazy lookups (`lazyPreviewSequence`, `Sequence<CollectedPreview>`,
factory-lambda type) are owned by `PreviewSequenceBuildContext` so all
sub-builders share one instance per IR transformer.

Test imports: `HintHashCollisionTest` now imports
`buildHashMapWithCollisionDetection` from the new logic path.

Quality gates (all green):
- `./gradlew ktlintFormat ktlintCheck --continue` — EXIT 0
- `./gradlew :compiler-plugin:test --rerun-tasks --continue` — EXIT 0
- `./gradlew jvmTest --rerun-tasks --continue` — EXIT 0
- `./gradlew apiCheck --continue` — EXIT 0
